Update android ProgressListenerInterface.kt

+++ android/src/main/java/com/mattermost/networkclient/uploads/ProgressListenerInterface.kt
@@ -2,5 +2,5 @@ package com.mattermost.networkclient.uploads
 
 interface ProgressListenerInterface {
     fun update(bytesRead: Long, contentLength: Long)
-    fun emitProgressEvent(progress: Int)
+    fun emitProgressEvent(progress: Long)
 }
